fix(config): set config vercel again
